backend/requirements.txt
fastapi 
uvicorn 
pymongo 
python-dotenv 
fastapi-mail 
werkzeug
pydantic
passlib[bcrypt]
PyJWT
requests
python-whois
bs4
numpy
xgboost
pandas
beautifulsoup4
googlesearch-python
joblib
tld
scikit-learn
matplotlib
python-multipart